% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmcDementiaPredictionTable1.R
\docType{package}
\name{EmcDementiaPredictionTable1}
\alias{EmcDementiaPredictionTable1}
\title{EmcDementiaPredictionTable1}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
